after_success: refactor, refactor, refactor!
Well, that was fun.

So, instead of messing around with the working tree, creating an orphan branch
if one can't be found, etc, etc, we simply go and construct our tree manually.
We use git-write-commit to build out a commit using that tree; if we can find
an existing remote branch (via git fetch) we add it as a parent.

Then *that* commit gets pushed back to github, to the target branch's ref.
whew!